+++ crates/core/src/config/mod.rs
@@ -65,6 +65,18 @@ pub struct Config {
 
     #[serde(default)]
     pub compression_format: CompressionFormat,
+
+    #[serde(default)]
+    #[cfg(target_os = "windows")]
+    pub close_to_tray: bool,
+
+    #[serde(default)]
+    #[cfg(target_os = "windows")]
+    pub autostart: bool,
+
+    #[serde(default)]
+    #[cfg(target_os = "windows")]
+    pub start_closed_to_tray: bool,
 }
 
 impl Config {