+++ src/main.rs
@@ -7,8 +7,11 @@ mod cli;
 mod command;
 mod gui;
 mod localization;
+#[cfg(target_os = "windows")]
+mod tray;
 
-use ajour_core::fs::CONFIG_DIR;
+use ajour_core::config::Config;
+use ajour_core::fs::{PersistentData, CONFIG_DIR};
 use ajour_core::utility::{remove_file, rename};
 
 #[cfg(target_os = "linux")]
@@ -87,8 +90,14 @@ pub fn main() {
             }
         }
         None => {
+            let config: Config =
+                Config::load_or_default().expect("loading config on application startup");
+
+            #[cfg(target_os = "windows")]
+            tray::spawn_sys_tray(config.close_to_tray, config.start_closed_to_tray);
+
             // Start the GUI
-            gui::run(opts);
+            gui::run(opts, config);
         }
     }
 }